Tried to avoid duplicating code, but it is unavoidable.
https://code.google.com/p/android/issues/detail?id=200941

To work around this UsefulDocumentFile is now a standalone class that implements duplicates of DocumentsContractApi19 & 21.  Most of the code is an single elegant type-agnostic implementation of the various DocumentFile subclasses: SingleDocumentFile, TreeDocumentFile, RawDocumentFile.

This is important because it allows us to retain tree uris and recreate the appropriate 'DocumentFile' representation.  We can now traverse tree uris in both directions (at least for filesystem-backed documents).